requirements.txt
moviepy==2.2.1
boto3==1.35.49
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
Pillow==11.3.0
numpy==2.0.2
imageio[ffmpeg]
